[CALCITE-3085] Add RelBaseShuttle, implementation of RelShuttle without stack
org.apache.calcite.rel.RelShuttleImpl class has a protected stack field
used to capture parent node when visiting children, but all subclasses
in Calcite codebase never read the content of the stack.

The change add a new implementation of RelShuttle named RelBasicShuttle.
The implementation do not use/populate a stack.

Also make the class the base class of RelShuttleImpl, along with all the
sub-classes of RelShuttleImpl (since none of them requires a stack).
