+++ core/src/main/java/org/apache/calcite/rel/RelShuttleImpl.java
@@ -16,119 +16,27 @@
  */
 package org.apache.calcite.rel;
 
-import org.apache.calcite.linq4j.Ord;
-import org.apache.calcite.rel.core.TableFunctionScan;
-import org.apache.calcite.rel.core.TableScan;
-import org.apache.calcite.rel.logical.LogicalAggregate;
-import org.apache.calcite.rel.logical.LogicalCorrelate;
-import org.apache.calcite.rel.logical.LogicalExchange;
-import org.apache.calcite.rel.logical.LogicalFilter;
-import org.apache.calcite.rel.logical.LogicalIntersect;
-import org.apache.calcite.rel.logical.LogicalJoin;
-import org.apache.calcite.rel.logical.LogicalMatch;
-import org.apache.calcite.rel.logical.LogicalMinus;
-import org.apache.calcite.rel.logical.LogicalProject;
-import org.apache.calcite.rel.logical.LogicalSort;
-import org.apache.calcite.rel.logical.LogicalUnion;
-import org.apache.calcite.rel.logical.LogicalValues;
-
 import java.util.ArrayDeque;
-import java.util.ArrayList;
 import java.util.Deque;
-import java.util.List;
 
 /**
- * Basic implementation of {@link RelShuttle} that calls
+ * Implementation of {@link RelShuttle} that calls
  * {@link RelNode#accept(RelShuttle)} on each child, and
  * {@link RelNode#copy(org.apache.calcite.plan.RelTraitSet, java.util.List)} if
- * any children change.
+ * any children change. It also keeps track of parents by adding them to a
+ * protected stack before visiting the child node. When the stack is not required,
+ * user should consider using {@link RelBasicShuttle} instead.
  */
-public class RelShuttleImpl implements RelShuttle {
+public class RelShuttleImpl extends RelBasicShuttle {
   protected final Deque<RelNode> stack = new ArrayDeque<>();
 
-  /**
-   * Visits a particular child of a parent.
-   */
-  protected RelNode visitChild(RelNode parent, int i, RelNode child) {
+  @Override protected RelNode visitChild(RelNode parent, int i, RelNode child) {
     stack.push(parent);
     try {
-      RelNode child2 = child.accept(this);
-      if (child2 != child) {
-        final List<RelNode> newInputs = new ArrayList<>(parent.getInputs());
-        newInputs.set(i, child2);
-        return parent.copy(parent.getTraitSet(), newInputs);
-      }
-      return parent;
+      return super.visitChild(parent, i, child);
     } finally {
       stack.pop();
     }
-  }
-
-  protected RelNode visitChildren(RelNode rel) {
-    for (Ord<RelNode> input : Ord.zip(rel.getInputs())) {
-      rel = visitChild(rel, input.i, input.e);
-    }
-    return rel;
-  }
-
-  public RelNode visit(LogicalAggregate aggregate) {
-    return visitChild(aggregate, 0, aggregate.getInput());
-  }
-
-  public RelNode visit(LogicalMatch match) {
-    return visitChild(match, 0, match.getInput());
-  }
-
-  public RelNode visit(TableScan scan) {
-    return scan;
-  }
-
-  public RelNode visit(TableFunctionScan scan) {
-    return visitChildren(scan);
-  }
-
-  public RelNode visit(LogicalValues values) {
-    return values;
-  }
-
-  public RelNode visit(LogicalFilter filter) {
-    return visitChild(filter, 0, filter.getInput());
-  }
-
-  public RelNode visit(LogicalProject project) {
-    return visitChild(project, 0, project.getInput());
-  }
-
-  public RelNode visit(LogicalJoin join) {
-    return visitChildren(join);
-  }
-
-  public RelNode visit(LogicalCorrelate correlate) {
-    return visitChildren(correlate);
-  }
-
-  public RelNode visit(LogicalUnion union) {
-    return visitChildren(union);
-  }
-
-  public RelNode visit(LogicalIntersect intersect) {
-    return visitChildren(intersect);
-  }
-
-  public RelNode visit(LogicalMinus minus) {
-    return visitChildren(minus);
-  }
-
-  public RelNode visit(LogicalSort sort) {
-    return visitChildren(sort);
-  }
-
-  public RelNode visit(LogicalExchange exchange) {
-    return visitChildren(exchange);
-  }
-
-  public RelNode visit(RelNode other) {
-    return visitChildren(other);
   }
 }
